requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
cytoolz==0.10.1
ECPy==1.2.3
eth-hash==0.2.0
eth-keys==0.3.3
eth-typing==2.2.1
eth-utils==1.8.4
future==0.18.2
hidapi==0.9.0.post2
idna==2.9
ledgerblue==0.1.31
Pillow==7.1.2
protobuf==3.11.3
pycryptodome==3.9.7
pycryptodomex==3.9.7
pysha3==1.0.2
python-u2flib-host==3.0.3
requests==2.23.0
rlp==1.2.0
six==1.14.0
toolz==0.10.0
urllib3==1.25.9
websocket-client==0.57.0
